Allow up to 2^31 constants per chunk
This is kind of an overkill, the 2^24 limit we had before was large
enough already. But this change is good because it simplifies things a
little bit because we can use Go's standard library for encoding and
decoding integers. This will be even more important for upcoming
instructions that will use multi-byte signed integer operands.