Make VM.runtimeError() panic, use it instead
Previously I was calling `panic()` explicitly to report some runtime
errors. I think initially I had the idea of using `panic()` for those
errors that were supposed to never happen (unless there was a bug in the
code generation) and `runtimeError()` for the "possible" runtime errors.

Anyway, given that

(1) the current usage was not consistent
(2) I need to revisit runtime error handling at some point
(3) ideally we should not have runtime errors, ever

I thought that, for now, it would be better to use `runtimeError()`
everywhere.